// src/services/dashboard/type.ts
export interface IBaseName {
  code: string;
  name: string;
}
// ...
export interface ISubject extends IBaseName {
  id: number;
}
// ...
export interface ISubjects {
  subject: {
    id: number;
    code: string;
    name: string;
  };
  subjectType: {
    code: string;
    name: string;
  };
  _semester: number;
  total_acload: number;
  credit: number;
}
// ...
export interface ICurriculumSubject {
  subject: ISubject;
  subjectType: IBaseName;
  _semester: string;
  total_acload: number;
  credit: number;
}
// ...
export enum MarkingSystemCodeEnum {
  Rating = '11',
  Grade = '12',
  Credit = '13',
}